% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\docType{data}
\name{cache}
\alias{cache}
\title{Environment that we use to cache the DuckDB connections}
\format{
An object of class \code{environment} of length 1.
}
\usage{
cache
}
\description{
Environment that we use to cache the DuckDB connections
}
\keyword{datasets}
