% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{clear_cache}
\alias{clear_cache}
\title{Clear the default cache directory}
\usage{
clear_cache()
}
\value{
A length one character vector.
}
\description{
Clear the default cache directory
}
\keyword{internal}
