% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/counts.R
\docType{data}
\name{COUNTS_URL}
\alias{COUNTS_URL}
\title{Base URL pointing to the count data}
\format{
An object of class \code{character} of length 1.
}
\usage{
COUNTS_URL
}
\description{
Base URL pointing to the count data
}
\keyword{datasets}
