% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/counts.R
\docType{data}
\name{COUNTS_VERSION}
\alias{COUNTS_VERSION}
\title{Current version of the counts. This will be incremented when a newer
version is released}
\format{
An object of class \code{character} of length 1.
}
\usage{
COUNTS_VERSION
}
\description{
Current version of the counts. This will be incremented when a newer
version is released
}
\keyword{datasets}
