% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\docType{data}
\name{get_database_url}
\alias{get_database_url}
\title{URL pointing to the full metadata file}
\format{
An object of class \code{character} of length 1.
}
\usage{
get_database_url
}
\value{
A character scalar consisting of the URL
}
\description{
URL pointing to the full metadata file
}
\examples{
get_metadata(remote_url = get_database_url())
}
\keyword{datasets}
