% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dev.R
\name{downsample_metadata}
\alias{downsample_metadata}
\title{Makes a "downsampled" metadata file that only contains the minimal data
needed to run the vignette.}
\usage{
downsample_metadata(output = "sample_meta.parquet")
}
\arguments{
\item{output}{Character scalar. Path to the output file.}
}
\description{
Makes a "downsampled" metadata file that only contains the minimal data
needed to run the vignette.
}
\keyword{internal}
