% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{get_database_url}
\alias{get_database_url}
\title{Returns the URLs for all metadata files}
\usage{
get_database_url(
  databases = c("metadata.0.2.3.parquet", "fibrosis.0.2.3.parquet")
)
}
\arguments{
\item{databases}{A character vector specifying the names of the metadata files. Default is c("metadata.0.2.3.parquet", "fibrosis.0.2.3.parquet")}
}
\value{
A character vector of URLs to parquet files to download
}
\description{
Returns the URLs for all metadata files
}
\examples{
get_database_url("metadata.0.2.3.parquet")
}
