% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{get_metadata}
\alias{get_metadata}
\title{Returns a data frame of Human Cell Atlas metadata, which should be filtered
and ultimately passed into get_SingleCellExperiment.}
\usage{
get_metadata(
  sqlite_path = "/vast/projects/RCP/human_cell_atlas/metadata.sqlite"
)
}
\arguments{
\item{sqlite_path}{Path to the sqlite database where the metadata can be found.
Currently this defaults to an internal location within WEHI's milton system.}
}
\value{
A lazy data.frame subclass containing the metadata. You can interact
with this object using most standard dplyr functions. However, it is recommended 
that you use the %LIKE% operator for string matching, as most stringr functions
will not work
}
\description{
Returns a data frame of Human Cell Atlas metadata, which should be filtered
and ultimately passed into get_SingleCellExperiment.
}
\examples{
filtered_metadata = get_metadata() |> 
  filter(
    ethnicity == "African" & 
    assay \%LIKE\% "\%10x\%" & 
    tissue == "lung parenchyma" &
    cell_type \%LIKE\% "\%CD4\%"
  )

}
