% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/counts.R
\name{get_single_cell_experiment}
\alias{get_single_cell_experiment}
\title{Gets a SingleCellExperiment from curated metadata}
\source{
\href{https://www.biorxiv.org/content/10.1101/2023.06.08.542671v3}{Mangiola et al.,2023}
}
\usage{
get_single_cell_experiment(data, ...)
}
\arguments{
\item{data}{A data frame containing, at minimum, \code{cell_}, \code{file_id_db} columns,
which correspond to a single cell ID, file subdivision for internal use.
They can be obtained from the \code{\link[=get_metadata]{get_metadata()}} function.}

\item{...}{
  Arguments passed on to \code{\link[=get_data_container]{get_data_container}}
  \describe{
    \item{\code{assays}}{A character vector whose elements must be either "counts"
and/or "cpm" and/or "quantile_normalised", representing the corresponding assay(s) you want to request.
By default only the count assay is downloaded. If you are interested in
comparing a limited amount of genes, the "cpm" assay is more appropriate.}
    \item{\code{cache_directory}}{An optional character vector of length one. If
provided, it should indicate a local file path where any remotely accessed
files should be copied.}
    \item{\code{repository}}{A character vector of length one. If provided, it should be
an HTTP URL pointing to the location where the single cell data is stored.}
    \item{\code{grouping_column}}{A character vector of metadata column for grouping. "file_id_db" for
for Single Cell Experiment, "file_id" for pseudobulk.}
    \item{\code{features}}{An optional character vector of features (ie genes) to return
the counts for. By default counts for all features will be returned.}
  }}
}
\description{
Gets a SingleCellExperiment from curated metadata
}
\examples{
meta <- get_metadata() |> head(2)
sce <- get_single_cell_experiment(meta)
}
\references{
Mangiola, S., M. Milton, N. Ranathunga, C. S. N. Li-Wai-Suen,
A. Odainic, E. Yang, W. Hutchison et al. "A multi-organ map of the human
immune system across age, sex and ethnicity." bioRxiv (2023): 2023-06.
doi:10.1101/2023.06.08.542671.
}
