% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{get_SingleCellExperiment}
\alias{get_SingleCellExperiment}
\title{Given a data frame of HCA metadata, returns a SingleCellExperiment object corresponding to the samples in that data frame}
\usage{
get_SingleCellExperiment(
  .data,
  assay = "counts",
  repository = when(assay, equals(., "counts") ~
    "/vast/projects/RCP/human_cell_atlas/splitted_DB2_data", equals(.,
    "counts_per_million") ~
    "/vast/projects/RCP/human_cell_atlas/splitted_DB2_data_scaled"),
  genes = NULL
)
}
\arguments{
\item{.data}{A data frame containing, at minimum, a `.sample` column, which corresponds to a single cell sample ID.
This can be obtained from the [get_metadata()] function.}

\item{assay}{TODO}

\item{repository}{A character vector of length one, which is a file path to where the single cell data is stored}

\item{genes}{An optional character vector of genes to return the counts for. By default counts for all genes will be returned.}
}
\description{
Given a data frame of HCA metadata, returns a SingleCellExperiment object corresponding to the samples in that data frame
}
