% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{get_unharmonised_metadata}
\alias{get_unharmonised_metadata}
\title{Returns unharmonised metadata for selected datasets.}
\usage{
get_unharmonised_metadata(
  dataset_ids,
  remote_url =
    "https://object-store.rc.nectar.org.au/v1/AUTH_06d6e008e3e642da99d806ba3ea629c5/unharmonised_metadata",
  cache_directory = get_default_cache_dir()
)
}
\arguments{
\item{dataset_ids}{A character vector, where each entry is a dataset ID
obtained from the \verb{$dataset_id} column of the table returned from
\code{\link[=get_metadata]{get_metadata()}}}
}
\value{
A named list, where each name is a dataset ID, and each value is
a "lazy data frame", ie a \code{tbl}.
}
\description{
Various metadata fields are \emph{not} common between datasets, so it does not make
sense for these to live in the main metadata table. This function is a
utility that allows easy fetching of this data if necessary.
}
\examples{
dataset_id = "838ea006-2369-4e2c-b426-b2a744a2b02b"
harmonised_meta = get_metadata() |> dplyr::filter(dataset_id = dataset_id)
unharmonised_meta = get_unharmonised_metadata(dataset_id)
unharmonised_tbl = unharmonised_meta[[dataset_id]]
dplyr::join(harmonised_meta, unharmonised_tbl, by=c("dataset_id", "cell_"))
}
