% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_metadata_and_counts.R
\name{import_metadata_counts}
\alias{import_metadata_counts}
\title{Checks importing criteria for new atlas
Generats counts per million from provided raw counts}
\usage{
import_metadata_counts(sce_rds = data, cache_dir = get_default_cache_dir())
}
\arguments{
\item{sce_rds}{A SingleCellExperiment object Rds}

\item{cache_dir}{Optional character vector of length 1. A file path on
your local system to a directory (not a file) that will be used to store
\code{metadata.parquet}}
}
\value{
A metadata.parquet, and a counts per million directory in provided cache directory
}
\description{
Checks importing criteria for new atlas
Generats counts per million from provided raw counts
}
\examples{
import_metadata_counts(sce_rds = "~/projects/caq/import_api_pipelines/12eb5fe25994253c1d320ca590a6e999/se.rds",
                       cache_directory = "~/projects/caq/cache_for_testing")

}
