% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{read_parquet}
\alias{read_parquet}
\title{Returns a tibble from a parquet file path}
\usage{
read_parquet(conn, path)
}
\value{
An SQL data frame
}
\description{
Since dbplyr 2.4.0, raw file paths aren't handled very well
See: \url{https://github.com/duckdb/duckdb-r/issues/38}
Hence the need for this method
}
\keyword{internal}
