% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sample_sce_obj}
\alias{sample_sce_obj}
\title{Sample SingleCellExperiment Object}
\format{
An object of class \code{SingleCellExperiment}
}
\source{
HeOrganAtlasData Liver
}
\usage{
sample_sce_obj
}
\description{
A sample SCE object from HeOrganAtlasData with transformation in metadata and assay
}
\keyword{datasets}
