% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{sync_remote_files}
\alias{sync_remote_files}
\title{Synchronises one or more remote files with a local copy}
\usage{
sync_remote_files(url, cache_dir, files)
}
\arguments{
\item{url}{A character vector of length one. The base HTTP url from which to obtain the files.}

\item{cache_dir}{A character vector of length one. The local filepath to synchronise files to.}

\item{files}{A character vector containing one or more file_id_db entries}
}
\value{
Subject to change
}
\description{
Synchronises one or more remote files with a local copy
}
