% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unharmonised.R
\docType{data}
\name{UNHARMONISED_URL}
\alias{UNHARMONISED_URL}
\title{Base URL for all the unharmonised data}
\format{
An object of class \code{character} of length 1.
}
\usage{
UNHARMONISED_URL
}
\description{
Base URL for all the unharmonised data
}
\keyword{datasets}
