% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\docType{data}
\name{DATABASE_URL}
\alias{DATABASE_URL}
\title{URL pointing to the full metadata file}
\format{
An object of class \code{character} of length 1.
}
\usage{
DATABASE_URL
}
\description{
URL pointing to the full metadata file
}
\keyword{datasets}
