% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/counts_per_million.R
\name{get_counts_per_million}
\alias{get_counts_per_million}
\title{Generating counts per million from SingleCellExperiment object}
\usage{
get_counts_per_million(input_sce_obj, output_dir, hd5_file_dir)
}
\arguments{
\item{input_sce_obj}{A SingleCellExperiment object read from RDS}

\item{output_dir}{A character vector of counts per million path}

\item{hd5_file_dir}{A character vector of HDF5 file path after converting from RDS}
}
\value{
A directory stores counts per million
}
\description{
Generating counts per million from SingleCellExperiment object
}
\examples{
\dontrun{
input_sce_obj <- readRDS("~/path/sample_sce.rds")
output_dir <- "~/cache_directory/cpm/file_id_db"
hd5_file_dir <- "~/cache_directory/original/file_id_db"
get_counts_per_million(input_sce_obj, output_dir, hd5_file_dir)
}
}
