% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{get_metadata}
\alias{get_metadata}
\title{Returns a data frame of Human Cell Atlas metadata, which should be filtered
and ultimately passed into get_SingleCellExperiment.}
\usage{
get_metadata(
  connection = list(dbname = "metadata", host = "7b4abe7csjh.db.cloud.edu.au", port =
    "5432", password = "password", user = "public_access")
)
}
\arguments{
\item{connection}{Optional list of postgres connection parameters used to
connect to the metadata database. Possible parameters are described here:
https://rpostgres.r-dbi.org/reference/postgres.}
}
\value{
A lazy data.frame subclass containing the metadata. You can interact
  with this object using most standard dplyr functions. For string matching, 
  it is recommended that you use `stringr::str_like` to filter character
  columns, as `stringr::str_match` will not work.
}
\description{
Returns a data frame of Human Cell Atlas metadata, which should be filtered
and ultimately passed into get_SingleCellExperiment.
}
\examples{
library(dplyr)
library(stringr)
filtered_metadata <- get_metadata() |>
    filter(
        ethnicity == "African" &
        str_like(assay, "\%10x\%") &
        tissue == "lung parenchyma" &
        str_like(cell_type, "\%CD4\%")
    )

}
