% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{get_metadata}
\alias{get_metadata}
\title{Returns a data frame of Human Cell Atlas metadata, which should be filtered
and ultimately passed into get_SingleCellExperiment.}
\usage{
get_metadata(sqlite_path = NULL)
}
\arguments{
\item{sqlite_path}{Path to the sqlite database where the metadata can be found.
Currently this defaults to an internal location within WEHI's milton system.}
}
\description{
Returns a data frame of Human Cell Atlas metadata, which should be filtered
and ultimately passed into get_SingleCellExperiment.
}
