% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/counts.R
\name{get_pseudobulk}
\alias{get_pseudobulk}
\title{Gets a Pseudobulk from curated metadata}
\usage{
get_pseudobulk(...)
}
\arguments{
\item{...}{
  Arguments passed on to \code{\link[=get_summarized_experiment]{get_summarized_experiment}}
  \describe{
    \item{\code{type}}{A character vector of length one. Either sce for Single Cell Experiment,
or pseudobulk.}
    \item{\code{data}}{A data frame containing, at minimum, \code{cell_}, \code{file_id_db}, \code{file_id} column, which
correspond to a single cell ID, file subdivision for internal use, and a single cell sample ID.
They can be obtained from the \code{\link[=get_metadata]{get_metadata()}} function.}
    \item{\code{assays}}{A character vector whose elements must be either "counts"
and/or "cpm" and/or "quantile_normalised", representing the corresponding assay(s) you want to request.
By default only the count assay is downloaded. If you are interested in
comparing a limited amount of genes, the "cpm" assay is more appropriate.}
    \item{\code{repository}}{A character vector of length one. If provided, it should be
an HTTP URL pointing to the location where the single cell data is stored.}
    \item{\code{cache_directory}}{An optional character vector of length one. If
provided, it should indicate a local file path where any remotely accessed
files should be copied.}
    \item{\code{features}}{An optional character vector of features (ie genes) to return
the counts for. By default counts for all features will be returned.}
  }}
}
\description{
Gets a Pseudobulk from curated metadata
}
\examples{
\dontrun{
meta <- get_metadata() |> filter(tissue_harmonised == "lung")
pseudobulk <- meta |> get_pseudobulk()
}
}
