% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seurat.R
\name{get_seurat}
\alias{get_seurat}
\title{Given a data frame of HCA metadata, returns a Seurat object corresponding to
the samples in that data frame}
\usage{
get_seurat(...)
}
\arguments{
\item{...}{
  Arguments passed on to \code{\link[=get_single_cell_experiment]{get_single_cell_experiment}}
  \describe{
    \item{\code{data}}{A data frame containing, at minimum, \code{cell_}, \code{file_id_db} columns,
which correspond to a single cell ID, file subdivision for internal use.
They can be obtained from the \code{\link[=get_metadata]{get_metadata()}} function.}
  }}
}
\value{
A Seurat object containing the same data as a call to
\code{\link[=get_single_cell_experiment]{get_single_cell_experiment()}}
}
\description{
Given a data frame of HCA metadata, returns a Seurat object corresponding to
the samples in that data frame
}
\examples{
meta <- get_metadata() |> head(2)
seurat <- get_seurat(meta)

}
