% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seurat.R
\name{get_seurat}
\alias{get_seurat}
\title{Given a data frame of HCA metadata, returns a Seurat object corresponding to
the samples in that data frame}
\source{
\href{https://www.biorxiv.org/content/10.1101/2023.06.08.542671v3}{Mangiola et al.,2023}
}
\usage{
get_seurat(...)
}
\arguments{
\item{...}{
  Arguments passed on to \code{\link[=get_single_cell_experiment]{get_single_cell_experiment}}
  \describe{
    \item{\code{data}}{A data frame containing, at minimum, \code{cell_}, \code{file_id_db} columns,
which correspond to a single cell ID, file subdivision for internal use.
They can be obtained from the \code{\link[=get_metadata]{get_metadata()}} function.}
  }}
}
\value{
A Seurat object containing the same data as a call to
\code{\link[=get_single_cell_experiment]{get_single_cell_experiment()}}
}
\description{
Given a data frame of HCA metadata, returns a Seurat object corresponding to
the samples in that data frame
}
\examples{
meta <- get_metadata() |> head(2)
seurat <- get_seurat(meta)
}
\references{
Mangiola, S., M. Milton, N. Ranathunga, C. S. N. Li-Wai-Suen,
A. Odainic, E. Yang, W. Hutchison et al. "A multi-organ map of the human
immune system across age, sex and ethnicity." bioRxiv (2023): 2023-06.
doi:10.1101/2023.06.08.542671.
}
