% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{get_seurat}
\alias{get_seurat}
\title{Given a data frame of HCA metadata, returns a Seurat object corresponding to the samples in that data frame}
\usage{
get_seurat(...)
}
\arguments{
\item{...}{
  Arguments passed on to \code{\link[=get_SingleCellExperiment]{get_SingleCellExperiment}}
  \describe{
    \item{\code{.data}}{A data frame containing, at minimum, a `.sample` column, which corresponds to a single cell sample ID.
This can be obtained from the [get_metadata()] function.}
    \item{\code{repository}}{A character vector of length one, which is a file path to where the single cell data is stored}
    \item{\code{genes}}{An optional character vector of genes to return the counts for. By default counts for all genes will be returned.}
    \item{\code{assay}}{TODO}
  }}
}
\description{
Given a data frame of HCA metadata, returns a Seurat object corresponding to the samples in that data frame
}
