% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/counts.R
\name{get_single_cell_experiment}
\alias{get_single_cell_experiment}
\title{Gets a SingleCellExperiment from curated metadata}
\usage{
get_single_cell_experiment(
  data,
  assays = "counts",
  cache_directory = get_default_cache_dir(),
  repository = COUNTS_URL,
  features = NULL
)
}
\arguments{
\item{data}{A data frame containing, at minimum, \code{cell_}, \code{file_id_db}, \code{file_id} column, which
correspond to a single cell ID, file subdivision for internal use, and a single cell sample ID.
They can be obtained from the \code{\link[=get_metadata]{get_metadata()}} function.}

\item{assays}{A character vector whose elements must be either "counts"
and/or "cpm", representing the corresponding assay(s) you want to request.
By default only the count assay is downloaded. If you are interested in
comparing a limited amount of genes, the "cpm" assay is more appropriate.}

\item{cache_directory}{An optional character vector of length one. If
provided, it should indicate a local file path where any remotely accessed
files should be copied.}

\item{repository}{A character vector of length one. If provided, it should be
an HTTP URL pointing to the location where the single cell data is stored.}

\item{features}{An optional character vector of features (ie genes) to return
the counts for. By default counts for all features will be returned.}
}
\value{
A SingleCellExperiment object, with one assay for each value in the
assays argument
}
\description{
Given a data frame of Curated Atlas metadata obtained from \code{\link[=get_metadata]{get_metadata()}},
returns a \code{\link[SingleCellExperiment:SingleCellExperiment]{SingleCellExperiment::SingleCellExperiment}} object
corresponding to the samples in that data frame
}
\examples{
meta <- get_metadata() |> head(2)
sce <- get_single_cell_experiment(meta)

}
