% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/counts.R
\name{get_SingleCellExperiment}
\alias{get_SingleCellExperiment}
\title{Gets a SingleCellExperiment from curated metadata}
\source{
\href{https://www.biorxiv.org/content/10.1101/2023.06.08.542671v3}{Mangiola et al.,2023}
}
\usage{
get_SingleCellExperiment(...)
}
\arguments{
\item{...}{
  Arguments passed on to \code{\link[=get_single_cell_experiment]{get_single_cell_experiment}}
  \describe{
    \item{\code{data}}{A data frame containing, at minimum, \code{cell_}, \code{file_id_db} columns,
which correspond to a single cell ID, file subdivision for internal use.
They can be obtained from the \code{\link[=get_metadata]{get_metadata()}} function.}
  }}
}
\description{
Gets a SingleCellExperiment from curated metadata
}
\examples{
meta <- get_metadata() |> head(2)
sce <- get_single_cell_experiment(meta)
}
\references{
Mangiola, S., M. Milton, N. Ranathunga, C. S. N. Li-Wai-Suen,
A. Odainic, E. Yang, W. Hutchison et al. "A multi-organ map of the human
immune system across age, sex and ethnicity." bioRxiv (2023): 2023-06.
doi:10.1101/2023.06.08.542671.
}
