% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/counts.R
\name{get_SingleCellExperiment}
\alias{get_SingleCellExperiment}
\title{Gets a SingleCellExperiment from curated metadata}
\usage{
get_SingleCellExperiment(...)
}
\arguments{
\item{...}{
  Arguments passed on to \code{\link[=get_single_cell_experiment]{get_single_cell_experiment}}
  \describe{
    \item{\code{}}{}
  }}
}
\description{
Gets a SingleCellExperiment from curated metadata
}
\examples{
meta <- get_metadata() |> head(2)
sce <- get_single_cell_experiment(meta)
}
