% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{get_SingleCellExperiment}
\alias{get_SingleCellExperiment}
\title{Given a data frame of HCA metadata, returns a SingleCellExperiment object corresponding to the samples in that data frame}
\usage{
get_SingleCellExperiment(
  data,
  assays = c("counts", "cpm"),
  cache_directory = get_default_cache_dir(),
  repository = REMOTE_URL,
  features = NULL
)
}
\arguments{
\item{data}{A data frame containing, at minimum, a `.sample` column, which corresponds to a single cell sample ID.
This can be obtained from the [get_metadata()] function.}

\item{assays}{A character vector whose elements must be either "raw" or "scaled", representing the corresponding assay you want to request.}

\item{cache_directory}{An optional character vector of length one. If provided, it should indicate a local file path where any remotely accessed files should be copied.}

\item{repository}{A character vector of length one. If provided, it should be an HTTP URL pointing to the location where the single cell data is stored.}

\item{features}{An optional character vector of features (ie genes) to return the counts for. By default counts for all features will be returned.}
}
\value{
A SingleCellExperiment object, with one assay for each value in the assays argument
}
\description{
Given a data frame of HCA metadata, returns a SingleCellExperiment object corresponding to the samples in that data frame
}
\examples{
meta <- get_metadata() |> head(2)
sce <- get_SingleCellExperiment(meta, repository = Sys.getenv("REMOTE_HCA"), cache_directory)

}
