% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_metadata_and_counts.R
\name{import_metadata_counts}
\alias{import_metadata_counts}
\title{Checks importing criteria for new atlas
Generats counts per million from provided raw counts}
\usage{
import_metadata_counts(sce_obj, cache_dir = get_default_cache_dir())
}
\arguments{
\item{sce_obj}{A SingleCellExperiment object from RDS}

\item{cache_dir}{Optional character vector of length 1. A file path on
your local system to a directory (not a file) that will be used to store
\code{metadata.parquet}}
}
\value{
A metadata.parquet strip from SCE object.
Directories store counts and counts per million in the provided cache directory.
}
\description{
Checks importing criteria for new atlas
Generats counts per million from provided raw counts
}
\examples{
data(sample_sce_obj)
import_metadata_counts(sample_sce_obj,
                       cache_dir = get_default_cache_dir())
}
