% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_metadata_and_counts.R
\name{import_metadata_counts}
\alias{import_metadata_counts}
\title{Checks importing criteria for new atlas
Calculating counts per million from raw counts}
\usage{
import_metadata_counts(
  meta_input,
  cache_dir = CuratedAtlasQueryR:::get_default_cache_dir(),
  meta_output,
  version = "0.2.3"
)
}
\arguments{
\item{meta_input}{A tibble}

\item{cache_dir}{Optional character vector of length 1. A file path on
your local system to a directory (not a file) that will be used to store
\code{metadata.parquet}}

\item{meta_output}{A parquet}
}
\value{
A directory stores counts per million
}
\description{
Checks importing criteria for new atlas
Calculating counts per million from raw counts
}
\examples{
example_metadata <- get_metadata() |> head(3) |> as_tibble()
import_metadata_counts(meta_input = example_metadata,
                       meta_output = "example_metadata",
                       cache_dir = tempdir())

}
