% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_metadata_and_counts.R
\name{import_one_sce}
\alias{import_one_sce}
\title{Import and process metadata and counts for a SingleCellExperiment object}
\usage{
import_one_sce(sce_obj, cache_dir = get_default_cache_dir())
}
\arguments{
\item{sce_obj}{A SingleCellExperiment object from RDS, the metadata slot of which
must contain \code{cell_} and \code{dataset_id}}

\item{cache_dir}{Optional character vector of length 1. A file path on
your local system to a directory (not a file) that will be used to store
\code{metadata.parquet}}
}
\value{
A metadata.parquet strip from the SingleCellExperiment object.
Directories store counts and counts per million in the provided cache directory.
}
\description{
Import and process metadata and counts for a SingleCellExperiment object
}
\examples{
data(sample_sce_obj)
import_one_sce(sample_sce_obj,
               cache_dir = get_default_cache_dir())
}
