% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\docType{data}
\name{SAMPLE_DATABASE_URL}
\alias{SAMPLE_DATABASE_URL}
\title{URL pointing to the sample metadata file, which is smaller and for test,
demonstration, and vignette purposes only}
\format{
An object of class \code{character} of length 1.
}
\usage{
SAMPLE_DATABASE_URL
}
\value{
A character scalar consisting of the URL
}
\description{
URL pointing to the sample metadata file, which is smaller and for test,
demonstration, and vignette purposes only
}
\examples{
get_metadata(remote_url = SAMPLE_DATABASE_URL)
}
\keyword{datasets}
