% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sample_sce_obj}
\alias{sample_sce_obj}
\title{Sample SingleCellExperiment Object}
\format{
An object of class \code{SingleCellExperiment}
}
\source{
HeOrganAtlasData Liver

\href{https://www.biorxiv.org/content/10.1101/2023.06.08.542671v3}{Mangiola et al.,2023}
}
\usage{
sample_sce_obj
}
\description{
A sample SingleCellExperiment object from HeOrganAtlasData with transformation in metadata and assay
}
\references{
Mangiola, S., M. Milton, N. Ranathunga, C. S. N. Li-Wai-Suen,
A. Odainic, E. Yang, W. Hutchison et al. "A multi-organ map of the human
immune system across age, sex and ethnicity." bioRxiv (2023): 2023-06.
doi:10.1101/2023.06.08.542671.
}
\keyword{datasets}
